refactor: 클래스 책임에 맞게 패키지 구조 변경
