+++ src/main/java/codesquad/config/ObjectContainer.java
@@ -1,7 +1,7 @@
 package codesquad.config;
 
-import codesquad.domain.UserStorage;
-import codesquad.http.RequestHandler;
+import codesquad.storage.UserStorage;
+import codesquad.server.RequestHandler;
 import codesquad.usecase.StaticResourceMappingUsecase;
 import codesquad.usecase.UserUsecase;
