+++ src/main/java/codesquad/server/StaticFilePathManager.java
@@ -1,4 +1,4 @@
-package codesquad.config;
+package codesquad.server;
 
 import java.util.List;
